+++ lib/rundoc/code_command/website/visit.rb
@@ -2,13 +2,14 @@
 
 class Rundoc::CodeCommand::Website
   class Visit < Rundoc::CodeCommand
-    def initialize(name:, url: nil, scroll: nil, height: 720, width: 1024, visible: false)
+    def initialize(name:, url: nil, scroll: nil, height: 720, width: 1024, visible: false, max_attempts: 3)
       @name = name
       @url = url
       @scroll = scroll
       @height = height
       @width = width
       @visible = visible
+      @max_attempts = max_attempts
     end
 
     def driver
@@ -33,7 +34,7 @@ class Rundoc::CodeCommand::Website
 
       puts message
 
-      driver.visit(@url) if @url
+      driver.visit(@url, max_attempts: @max_attempts) if @url
       driver.scroll(@scroll) if @scroll
 
       return "" if contents.nil? || contents.empty?